Move generate-config from ctl to patroni
