+++ patroni/ctl.py
@@ -1415,279 +1415,3 @@ def format_pg_version(version: int) -> str:
         return "{0}.{1}.{2}".format(version // 10000, version // 100 % 100, version % 100)
     else:
         return "{0}.{1}".format(version // 10000, version % 100)
-
-
-def get_bin_dir_from_running_instance(data_dir: str) -> str:
-    postmaster_pid = None
-    try:
-        with open(f"{data_dir}/postmaster.pid", 'r') as f:
-            postmaster_pid = f.readline()
-            if not postmaster_pid:
-                raise PatroniCtlException('Failed to obtain postmaster pid from postmaster.pid file')
-            postmaster_pid = int(postmaster_pid.strip())
-    except OSError as e:
-        raise PatroniCtlException(f'Error while reading postmaster.pid file: {e}')
-    import psutil
-    try:
-        return os.path.dirname(psutil.Process(postmaster_pid).exe())
-    except psutil.NoSuchProcess:
-        raise PatroniCtlException('Obtained postmaster pid doesn\'t exist')
-
-
-def enrich_config_from_running_instance(config: Dict[str, Any], no_value_msg: str, dsn: Optional[str] = None) -> None:
-    """Get
-    - non-internal GUC values having configuration file, postmaster command line or environment variable as a source
-    - postgresql.connect_address, postgresql.listen,
-    - postgresql.pg_hba and postgresql.pg_ident
-    - superuser auth parameters (from the options used for connection)
-    And redefine scope with the clister_name GUC value if set
-
-    :param config: configuration parameters dict to be enriched
-    :param no_value_msg: str value to be used when a parameter value is not available
-    :param dsn: optional DSN string for the source running instance
-    """
-    from getpass import getuser
-    from patroni.postgresql.config import parse_dsn
-    from patroni.config import AUTH_ALLOWED_PARAMETERS_MAPPING
-
-    def get_local_ip() -> str:
-        patroni_socket = socket.socket(socket.AF_INET, socket.SOCK_DGRAM)
-        try:
-            patroni_socket.connect(('8.8.8.8', 80))
-            ip = patroni_socket.getsockname()[0]
-        except OSError as e:
-            raise PatroniCtlException(f'Failed to define local ip: {e}')
-        finally:
-            patroni_socket.close()
-        return ip
-
-    su_params: Dict[str, str] = {}
-    parsed_dsn = {}
-
-    if dsn:
-        parsed_dsn = parse_dsn(dsn)
-        if not parsed_dsn:
-            raise PatroniCtlException('Failed to parse DSN string')
-
-    # gather auth parameters for the superuser config
-    for conn_param, env_var in AUTH_ALLOWED_PARAMETERS_MAPPING.items():
-        val = parsed_dsn.get(conn_param, os.getenv(env_var))
-        if val:
-            su_params[conn_param] = val
-    # because we use "username" in the config for some reason
-    su_params['username'] = su_params.pop('user', getuser())
-    su_params['password'] = su_params.get('password') or click.prompt('Please enter the user password',
-                                                                      hide_input=True, default="")
-
-    from . import psycopg
-    try:
-        conn = psycopg.connect(dsn=dsn, password=su_params['password'])
-    except psycopg.Error as e:
-        raise PatroniCtlException(f'Failed to establish PostgreSQL connection: {e}')
-
-    with conn.cursor() as cur:
-        cur.execute("SELECT 1 FROM pg_roles WHERE rolname=%s AND rolsuper='t';", (su_params['username'],))
-        if cur.rowcount < 1:
-            raise PatroniCtlException('The provided user does not have superuser privilege')
-
-        cur.execute("SELECT name, current_setting(name) FROM pg_settings \
-                     WHERE context <> 'internal' \
-                     AND source IN ('configuration file', 'command line', 'environment variable') \
-                     AND category <> 'Write-Ahead Log / Recovery Target' \
-                     AND setting <> '(disabled)' \
-                     OR name IN ('hba_file', 'ident_file', 'config_file', \
-                                 'data_directory', \
-                                 'listen_addresses', 'port', \
-                                 'max_connections', 'max_worker_processes', 'max_wal_senders', \
-                                 'max_replication_slots', 'max_locks_per_transaction', 'max_prepared_transactions', \
-                                 'hot_standby', \
-                                 'wal_level', 'wal_log_hints', \
-                                 'wal_keep_segments', 'wal_keep_size', \
-                                 'track_commit_timestamp');")
-
-        # adjust values
-        for p, v in cur.fetchall():
-            if p == 'data_directory':
-                config['postgresql']['data_dir'] = v
-            elif p == 'cluster_name':
-                config['scope'] = v
-            elif p in ('archive_command', 'restore_command', 'archive_cleanup_command',
-                       'recovery_end_command', 'ssl_passphrase_command',
-                       'hba_file', 'ident_file', 'config_file'):
-                # write commands to the local config due to security implications
-                # write hba/ident/config_file to local config to ensure they are not removed later
-                if not config['postgresql']['parameters']:
-                    config['postgresql']['parameters'] = {}
-                config['postgresql']['parameters'][p] = v
-            else:
-                config['bootstrap']['dcs']['postgresql']['parameters'][p] = v
-
-    conn.close()
-
-    port = config['bootstrap']['dcs']['postgresql']['parameters']['port']
-    connect_host = parsed_dsn.get('host', os.getenv('PGHOST', get_local_ip()))
-    connect_port = parsed_dsn.get('port', os.getenv('PGPORT', port))
-    config['postgresql']['connect_address'] = f'{connect_host}:{connect_port}'
-    listen_addresses = config['bootstrap']['dcs']['postgresql']['parameters']['listen_addresses']
-    config['postgresql']['listen'] = f'{listen_addresses}:{port}'
-
-    try:
-        with open(f"{config['postgresql']['parameters']['hba_file']}", 'r') as f:
-            config['postgresql']['pg_hba'] = [i.strip() for i in f.readlines()
-                                              if i.startswith(('local',
-                                                               'host',
-                                                               'hostssl',
-                                                               'hostnossl',
-                                                               'hostgssenc',
-                                                               'hostnogssenc'))]
-    except OSError as e:
-        raise PatroniCtlException(f'Failed to read hba_file: {e}')
-
-    try:
-        with open(f"{config['postgresql']['parameters']['ident_file']}", 'r') as f:
-            config['postgresql']['pg_ident'] = [i.strip() for i in f.readlines() if i.strip() and not i.startswith('#')]
-    except OSError as e:
-        raise PatroniCtlException(f'Failed to read ident_file: {e}')
-
-    config['postgresql']['authentication'] = {
-        'superuser': su_params,
-        'replication': {'username': no_value_msg, 'password': no_value_msg}
-    }
-
-
-@ctl.command('generate-config',
-             help='Generate patroni sample configuration file or a configuration file for a running instance')
-@click.option('--scope', help='Scope parameter value')
-@click.option('--file', '-f', help='Full path to the configuration file to be created', default='/tmp/patroni.yml')
-@click.option('--sample', '-s', help='Create a sample config', is_flag=True)
-@click.option('--dsn', '-d',
-              help='Optional DSN string of the instance to be used as a source \
-                    of postgres configuration parameter values. Superuser connection is required.')
-@click.option('--bin-dir', '-b', help='Full path to the directory with PostgreSQL libraries')
-def generate_config(scope: str, file: str, sample: bool, dsn: Optional[str], bin_dir: Optional[str]) -> None:
-    """Generate Patroni configuration file
-
-    Gather all the available non-internal GUC values having configuration file, postmaster command line or environment
-    variable as a source and store them in the appropriate part of Patroni configuration (``postgresql.parameters`` or
-    ``bootsrtap.dcs.postgresql.parameters``). Either the provided DSN or PG env vars will be used for the connection.
-
-    The created configuration contains:
-    - ``scope``: the provided option value. Will be overwritten with the cluster_name GUC value if it is available
-    - ``name``: hostname
-    - ``bootsrtap.dcs``: section with all the parameters (incl. the majority of PG GUCs) set to their default values
-      defined by Patroni and adjusted by the source instances's configuration values.
-    - ``postgresql.parameters``: the source instance's archive_command, restore_command, archive_cleanup_command,
-      recovery_end_command, ssl_passphrase_command, hba_file, ident_file, config_file GUC values
-    - ``postgresql.bin_dir``: path to Postgres binaries gathered from the running instance or, if not available,
-      the value of the --bin-dir option if provided. Otherwise, an empty string.
-    - ``postgresql.datadir``: the value gathered from the corresponding PG GUC
-    - ``postgresql.listen``: source instance's listen_addresses and port GUC values
-    - ``postgresql.connect_address``: if possible, generated from the connection params
-    - ``postgresql.authentication``:
-        - superuser and replication users defined (if possible, usernames are set from the respective Patroni ENV vars,
-          otherwise the default 'postgres' and 'replicator' values are used).
-          If not a sample config, either DSN or PG ENV vars are used to define superuser authentication parameters.
-        - rewind user is defined for a sample config if PG version can be defined and PG version is 11+
-          (if possible, username is set from the respective Patroni ENV var)
-    - ``bootsrtap.dcs.postgresql.use_pg_rewind set to True if PG version is 11+
-    - ``postgresql.pg_hba`` defaults or the lines gathered from the source instance's hba_file
-    - ``postgresql.pg_ident`` the lines gathered from the source instance's ident_file
-
-    :param scope: Scope parameter value to write into the configuration.
-    :param file: Full path to the configuration file to be created (/tmp/patroni.yml by default).
-    :param sample: Optional flag. If set, no source instance will be used - generate config with some sane defaults.
-    :param dsn: Optional DSN string for the local instance to get GUC values from.
-    :param bin_dir: Optional path to Postgres binaries.
-    """
-    from patroni.config import Config
-    from patroni.validator import get_major_version
-
-    no_value_msg = '#FIXME'
-    pg_version = None
-
-    dynamic_config = Config.get_default_config()
-    dynamic_config['postgresql']['parameters'] = dict(dynamic_config['postgresql']['parameters'])
-    config: Dict[str, Any] = {
-        'scope': scope or no_value_msg,
-        'name': socket.gethostname(),
-        'bootstrap': {
-            'dcs': dynamic_config
-        },
-        'postgresql': {
-            'parameters': None,
-            'connect_address': no_value_msg,
-            'listen': no_value_msg,
-        },
-    }
-
-    if not sample:
-        enrich_config_from_running_instance(config, no_value_msg, dsn)
-
-    # some work around bin_dir definition
-    if sample and bin_dir:
-        config['postgresql']['bin_dir'] = bin_dir
-    elif not sample:
-        config['postgresql']['bin_dir'] = get_bin_dir_from_running_instance(config['postgresql']['data_dir'])
-
-    if 'bin_dir' in config['postgresql']:
-        # obtain version from the binary
-        try:
-            pg_version = postgres_major_version_to_int(get_major_version(config['postgresql']['bin_dir'] or None))
-        except PatroniException as e:
-            raise PatroniCtlException(str(e))
-    else:
-        config['postgresql']['bin_dir'] = ''
-
-    # generate sample config
-    if sample:
-        # some sane defaults or values set via Patroni env vars
-        replicator = os.getenv('PATRONI_REPLICATION_USERNAME', 'replicator')
-        config['postgresql']['authentication'] = {
-            'superuser': {
-                'username': os.getenv('PATRONI_SUPERUSER_USERNAME', 'postgres'),
-                'password': os.getenv('PATRONI_SUPERUSER_PASSWORD', no_value_msg)
-            },
-            'replication': {
-                'username': replicator,
-                'password': os.getenv('PATRONI_REPLICATION_PASSWORD', no_value_msg)
-            }
-        }
-
-        auth_method = 'scram-sha-256' if pg_version and pg_version >= 100000 else 'md5'
-        config['postgresql']['pg_hba'] = [
-            f'host all all all {auth_method}',
-            f'host replication {replicator} all {auth_method}'
-        ]
-
-        # add version-specific configuration
-        if pg_version:
-            from patroni.postgresql.config import ConfigHandler
-            wal_keep_param = 'wal_keep_segments' if pg_version < 130000 else 'wal_keep_size'
-            config['bootstrap']['dcs']['postgresql']['parameters'][wal_keep_param] =\
-                ConfigHandler.CMDLINE_OPTIONS[wal_keep_param][0]
-
-            if pg_version >= 110000:
-                config['bootstrap']['dcs']['postgresql']['use_pg_rewind'] = True
-                config['postgresql']['authentication']['rewind'] = {
-                    'username': os.getenv('PATRONI_REWIND_USERNAME', 'rewind_user'),
-                    'password': no_value_msg
-                }
-
-    # redundant values from the default config
-    del config['bootstrap']['dcs']['postgresql']['parameters']['listen_addresses']
-    del config['bootstrap']['dcs']['postgresql']['parameters']['port']
-    del config['bootstrap']['dcs']['postgresql']['parameters']['cluster_name']
-    del config['bootstrap']['dcs']['standby_cluster']
-
-    # no value instead of 'none' in the parsed yaml
-    yaml.add_representer(
-        type(None),
-        lambda dumper, _:
-            dumper.represent_scalar(u'tag:yaml.org,2002:null', ''),  # pyright: ignore [reportUnknownMemberType]
-        Dumper=yaml.SafeDumper
-    )
-    dir_path = os.path.dirname(file)
-    if dir_path and not os.path.isdir(dir_path):
-        os.makedirs(dir_path)
-    with open(file, 'w') as fd:
-        yaml.safe_dump(config, fd, default_flow_style=False)
