+++ tests/test_ctl.py
@@ -689,237 +689,6 @@ class TestCtl(unittest.TestCase):
         self.assertIn("Waiting for reinitialize to complete on: other", result.output)
         self.assertIn("Reinitialize is completed on: other", result.output)
 
-    @patch('patroni.psycopg.connect', psycopg_connect)
-    @patch('builtins.open', MagicMock())
-    @patch('socket.gethostname', Mock(return_value='testhost'))
-    @patch('os.makedirs')
-    @patch('yaml.safe_dump')
-    def test_generate_config(self, mock_config_dump, mock_makedir):
-        scope = 'sample_config'
-        self.maxDiff = None
-        # 1. Wrong input
-        # 1.1 Wrong DSN format
-        result = self.runner.invoke(ctl, ['generate-config',
-                                          '--scope', scope,
-                                          '--dsn', 'host:foo port:bar user:foobar'])
-        assert result.exit_code == 1
-
-        # 1.2 User is not a superuser
-        with patch.object(MockCursor, 'rowcount', PropertyMock(return_value=0), create=True):
-            result = self.runner.invoke(ctl, ['generate-config',
-                                              '--scope', scope,
-                                              '--dsn', 'host=foo port=bar user=foobar'])
-            assert result.exit_code == 1
-
-        no_value_msg = '#FIXME'
-        dynamic_config = Config.get_default_config()
-        dynamic_config['postgresql']['parameters'] = dict(dynamic_config['postgresql']['parameters'])
-        del dynamic_config['standby_cluster']
-        for p in ('cluster_name', 'listen_addresses', 'port'):
-            del dynamic_config['postgresql']['parameters'][p]
-
-        config = {
-            'scope': scope,
-            'name': 'testhost',
-            'bootstrap': {
-                'dcs': dynamic_config
-            },
-            'postgresql': {
-                'connect_address': no_value_msg,
-                'listen': no_value_msg,
-                'pg_hba': ['host all all all md5', 'host replication replicator all md5'],
-                'parameters': None,
-                'authentication': {'superuser': {'username': 'postgres', 'password': no_value_msg},
-                                   'replication': {'username': 'replicator', 'password': no_value_msg}},
-                'bin_dir': ''
-            }
-        }
-
-        # 2. Sample config
-        # 2.1 With a dir creation
-        self.runner.invoke(ctl, ['generate-config', '-s', '--scope', scope, '--file', '/foo/bar.yml'])
-        mock_makedir.assert_called_once()
-        self.assertEqual(config, mock_config_dump.call_args[0][0])
-
-        mock_makedir.reset_mock()
-        mock_config_dump.reset_mock()
-
-        # 2.2 With bin_dir provided
-        config['postgresql']['bin_dir'] = '/foo/bar'
-
-        # Version-specific params
-
-        # 2.2.1 pg_version < 13
-        config['bootstrap']['dcs']['postgresql']['parameters']['wal_keep_segments'] = 8
-        with patch('subprocess.check_output', Mock(return_value=b"postgres (PostgreSQL) 9.4.3")):
-            self.runner.invoke(ctl, ['generate-config', '-s', '--scope', scope, '--bin-dir', '/foo/bar'])
-            mock_makedir.assert_not_called()
-            self.assertEqual(config, mock_config_dump.call_args[0][0])
-
-        mock_makedir.reset_mock()
-        mock_config_dump.reset_mock()
-
-        # 2.2.1 pg_version >= 13
-        del config['bootstrap']['dcs']['postgresql']['parameters']['wal_keep_segments']
-        config['bootstrap']['dcs']['postgresql']['parameters']['wal_keep_size'] = '128MB'
-        config['bootstrap']['dcs']['postgresql']['use_pg_rewind'] = True
-        config['postgresql']['authentication']['rewind'] = {'username': 'rewind_user', 'password': no_value_msg}
-        config['postgresql']['pg_hba'] = ['host all all all scram-sha-256',
-                                          'host replication replicator all scram-sha-256']
-        with patch('subprocess.check_output', Mock(return_value=b"postgres (PostgreSQL) 15.2")):
-            self.runner.invoke(ctl, ['generate-config', '-s', '--scope', scope, '--bin-dir', '/foo/bar'])
-            self.assertEqual(config, mock_config_dump.call_args[0][0])
-
-        mock_config_dump.reset_mock()
-
-        # 3. Config for a running instance (adjusted values are taken from tests/__init__.py)
-        config['scope'] = 'my_cluster'
-        config['postgresql']['connect_address'] = 'foo:bar'
-        config['postgresql']['listen'] = '6.6.6.6:1984'
-        config['postgresql']['parameters'] = {'archive_command': 'my archive command'}
-        config['postgresql']['parameters']['hba_file'] = '/hba/file/path'
-        config['postgresql']['parameters']['ident_file'] = '/ident/file/path'
-        config['postgresql']['pg_ident'] = ['foo bar foobar', 'foobar bar foo']
-        config['bootstrap']['dcs']['postgresql']['parameters']['max_connections'] = 42
-        config['bootstrap']['dcs']['postgresql']['parameters']['max_locks_per_transaction'] = 73
-        config['bootstrap']['dcs']['postgresql']['parameters']['max_replication_slots'] = 21
-        config['bootstrap']['dcs']['postgresql']['parameters']['max_wal_senders'] = 37
-        config['bootstrap']['dcs']['postgresql']['parameters']['wal_level'] = 'replica'
-
-        config['postgresql']['data_dir'] = '/foo/bar/data'
-        config['postgresql']['authentication']['superuser'] = {
-            'username': 'foobar',
-            'password': 'qwerty',
-            'channel_binding': 'prefer',
-            'gssencmode': 'prefer',
-            'sslmode': 'prefer'
-        }
-
-        config['postgresql']['authentication']['replication'] = {'username': no_value_msg, 'password': no_value_msg}
-        del config['bootstrap']['dcs']['postgresql']['use_pg_rewind']
-        del config['postgresql']['authentication']['rewind']
-        del config['bootstrap']['dcs']['postgresql']['parameters']['wal_keep_size']
-
-        hba_content = '\n'.join(config['postgresql']['pg_hba'] + ['#host all all all md5'])
-        ident_content = '\n'.join(config['postgresql']['pg_ident'] + ['# something very interesting', '  '])
-        open_res = []
-        for _ in range(5):
-            open_res.extend([
-                mock_open(read_data=hba_content)(),
-                mock_open(read_data=ident_content)(),
-                mock_open(read_data='1984')(),
-                mock_open()()
-            ])
-
-        # 3.1 bin_dir extracted from the running process
-        with patch('psutil.Process.__init__', Mock(return_value=None)),\
-             patch('psutil.Process.exe', Mock(return_value='/foo/bar/postgres')),\
-             patch('builtins.open', Mock(side_effect=open_res)):
-
-            # 3.1.1 explicitly provided bin_dir is different from the one extracted from the pid
-            # + PG version < 13
-            with patch('subprocess.check_output', Mock(return_value=b"postgres (PostgreSQL) 9.4.3")):
-                self.runner.invoke(ctl, ['generate-config', '--scope', scope, '--dsn',
-                                         'host=foo port=bar user=foobar password=qwerty',
-                                         '--bin-dir', '/bar/foo'])
-                self.assertEqual(config, mock_config_dump.call_args[0][0])
-
-                mock_config_dump.reset_mock()
-
-            # 3.1.2 pg_version >= 13
-            with patch('subprocess.check_output', Mock(return_value=b"postgres (PostgreSQL) 15.2")):
-                self.runner.invoke(ctl, ['generate-config', '--scope', scope, '--dsn',
-                                         'host=foo port=bar user=foobar password=qwerty'])
-                self.assertEqual(config, mock_config_dump.call_args[0][0])
-
-                mock_config_dump.reset_mock()
-
-            # 3.1.3 su auth params and connect host from env
-            with patch('subprocess.check_output', Mock(return_value=b"postgres (PostgreSQL) 15.2")):
-                os.environ['PGUSER'] = config['postgresql']['authentication']['superuser']['username'] = 'env_user'
-                os.environ['PGPASSWORD'] = config['postgresql']['authentication']['superuser']['password'] = 'env_pwd'
-                os.environ['PGCHANNELBINDING'] = \
-                    config['postgresql']['authentication']['superuser']['channel_binding'] = 'disable'
-                del config['postgresql']['authentication']['superuser']['gssencmode']
-                del config['postgresql']['authentication']['superuser']['sslmode']
-                os.environ['PGHOST'] = 'env_host'
-                config['postgresql']['connect_address'] = f"{os.environ['PGHOST']}:1984"
-
-                self.runner.invoke(ctl, ['generate-config', '--scope', scope])
-                self.assertEqual(config, mock_config_dump.call_args[0][0])
-
-                mock_config_dump.reset_mock()
-
-            with patch('subprocess.check_output', Mock(return_value=b"postgres (PostgreSQL) 15.2")):
-                # 3.1.4 su password from prompt
-                os.environ.pop('PGPASSWORD')
-                config['postgresql']['authentication']['superuser']['password'] = 'prompt_pwd'
-                self.runner.invoke(ctl, ['generate-config', '--scope', scope],
-                                   input=f"{config['postgresql']['authentication']['superuser']['password']}\n")
-                self.assertEqual(config, mock_config_dump.call_args[0][0])
-
-                mock_config_dump.reset_mock()
-
-            # 3.1.5 Error while calling postgres --version
-            with patch('subprocess.check_output', Mock(side_effect=OSError)):
-                result = self.runner.invoke(ctl, ['generate-config', '--scope', scope, '--dsn',
-                                                  'host=foo port=bar user=foobar password=qwerty',
-                                                  '--bin-dir', '/foo/bar'])
-                assert result.exit_code == 1
-
-        with patch('builtins.open', Mock(side_effect=[mock_open(read_data=hba_content)(),
-                                                      mock_open(read_data=ident_content)(),
-                                                      mock_open(read_data='')(),  # 3.2 empty postmaster.pid
-                                                      mock_open(read_data=hba_content)(),
-                                                      mock_open(read_data=ident_content)(),
-                                                      OSError,  # 3.3 Failed to open postmaster.pid
-                                                      mock_open(read_data=hba_content)(),
-                                                      mock_open(read_data=ident_content)(),
-                                                      mock_open(read_data='1984')()  # 3.4 Invalid postmaster pid
-                                                      ])):
-            result = self.runner.invoke(ctl, ['generate-config',
-                                              '--scope', scope,
-                                              '--dsn', 'host=foo port=bar user=foobar password=qwerty'])
-            assert result.exit_code == 1
-
-            result = self.runner.invoke(ctl, ['generate-config',
-                                              '--scope', scope,
-                                              '--dsn', 'host=foo port=bar user=foobar password=qwerty'])
-            assert result.exit_code == 1
-
-            with patch('psutil.Process.__init__', Mock(return_value=None)),\
-                 patch('psutil.Process.exe', Mock(side_effect=psutil.NoSuchProcess(1984))):
-                result = self.runner.invoke(ctl, ['generate-config',
-                                                  '--scope', scope,
-                                                  '--dsn', 'host=foo port=bar user=foobar, password=qwerty'])
-                assert result.exit_code == 1
-
-        # 3.5 Failed to open pg_hba
-        with patch('builtins.open', Mock(side_effect=OSError)):
-            result = self.runner.invoke(ctl, ['generate-config',
-                                              '--scope', scope,
-                                              '--dsn', 'host=foo port=bar user=foobar password=qwerty'])
-            assert result.exit_code == 1
-
-        # 3.6 Failed to open pg_ident
-        with patch('builtins.open', Mock(side_effect=[mock_open(read_data=hba_content)(), OSError])):
-            result = self.runner.invoke(ctl, ['generate-config',
-                                              '--scope', scope,
-                                              '--dsn', 'host=foo port=bar user=foobar password=qwerty'])
-            assert result.exit_code == 1
-
-        # 3.7 Failed PG connecttion
-        from . import psycopg
-        with patch('patroni.psycopg.connect', side_effect=psycopg.Error):
-            result = self.runner.invoke(ctl, ['generate-config', '--scope', scope])
-            assert result.exit_code == 1
-
-        # 3.8 Failed to get local IP
-        with patch('socket.socket.connect', Mock(side_effect=OSError)):
-            result = self.runner.invoke(ctl, ['generate-config', '--scope', scope])
-            assert result.exit_code == 1
-
-
 class TestPatronictlPrettyTable(unittest.TestCase):
 
     def setUp(self):
