+++ tests/test_patroni.py
@@ -1,17 +1,21 @@
 import etcd
+from io import StringIO
 import logging
 import os
+import psutil
 import signal
 import time
 import unittest
 
+from . import MockCursor
 import patroni.config as config
 from http.server import HTTPServer
-from mock import Mock, PropertyMock, patch
+from mock import MagicMock, Mock, PropertyMock, mock_open, patch
 from patroni.api import RestApiServer
 from patroni.async_executor import AsyncExecutor
 from patroni.dcs.etcd import AbstractEtcdClientWithFailover
-from patroni.exceptions import DCSError
+from patroni.config import Config
+from patroni.exceptions import DCSError, PatroniException
 from patroni.postgresql import Postgresql
 from patroni.postgresql.config import ConfigHandler
 from patroni import check_psycopg
@@ -60,6 +64,236 @@ class TestPatroni(unittest.TestCase):
         self.assertRaises(SystemExit, patroni_main)
         with patch.object(config.Config, '__init__', Mock(return_value=None)):
             self.assertRaises(SystemExit, patroni_main)
+
+    @patch('patroni.psycopg.connect', psycopg_connect)
+    @patch('builtins.open', MagicMock())
+    @patch('socket.gethostname', Mock(return_value='testhost'))
+    @patch('os.makedirs')
+    @patch('yaml.safe_dump')
+    @patch('sys.stderr', new_callable=StringIO)
+    def test_generate_sample_config(self, mock_stderr, mock_config_dump, mock_makedir):
+        self.maxDiff = None
+        os.environ['PATRONI_SCOPE'] = 'scope_from_env'
+        os.environ['PATRONI_POSTGRESQL_BIN_DIR'] = '/bin/from/env'
+        os.environ['PATRONI_SUPERUSER_USERNAME'] = 'su_user_from_env'
+        os.environ['PATRONI_SUPERUSER_PASSWORD'] = 'su_pwd_from_env'
+        os.environ['PATRONI_REPLICATION_USERNAME'] = 'repl_user_from_env'
+        os.environ['PATRONI_REPLICATION_PASSWORD'] = 'repl_pwd_from_env'
+        os.environ['PATRONI_REWIND_USERNAME'] = 'rewind_user_from_env'
+        os.environ['PGUSER'] = 'pguser_from_env'
+        os.environ['PGPASSWORD'] = 'pguser_pwd_from_env'
+        os.environ['PGHOST'] = 'pghost_from_env'
+
+        # 1. Wrong input
+        # 1.1 Wrong DSN format
+        with patch('sys.argv', ['patroni.py', '--generate-config', '--dsn', 'host:foo port:bar user:foobar']),\
+             self.assertRaises(SystemExit) as e:
+            patroni_main()
+        self.assertEqual(e.exception.code, 1)
+        self.assertIn('Failed to parse DSN string', mock_stderr.getvalue())
+        mock_stderr.truncate(0)
+        mock_stderr.seek(0)
+
+        # 1.2 User is not a superuser
+        with patch('sys.argv', ['patroni.py', '--generate-config', '--dsn', 'host=foo port=bar user=foobar password=pwd_from_dsn']),\
+             patch.object(MockCursor, 'rowcount', PropertyMock(return_value=0), create=True),\
+             self.assertRaises(SystemExit) as e:
+            patroni_main()
+        self.assertEqual(e.exception.code, 1)
+        self.assertIn('The provided user does not have superuser privilege', mock_stderr.getvalue())
+        mock_stderr.truncate(0)
+        mock_stderr.seek(0)
+
+        no_value_msg = '#FIXME'
+        dynamic_config = Config.get_default_config()
+        dynamic_config['postgresql']['parameters'] = dict(dynamic_config['postgresql']['parameters'])
+        del dynamic_config['standby_cluster']
+        for p in ('cluster_name', 'listen_addresses', 'port'):
+            del dynamic_config['postgresql']['parameters'][p]
+        dynamic_config['postgresql']['parameters']['wal_keep_segments'] = 8
+
+        config = {
+            'scope': os.environ['PATRONI_SCOPE'],
+            'name': 'testhost',
+            'bootstrap': {
+                'dcs': dynamic_config
+            },
+            'postgresql': {
+                'connect_address': no_value_msg,
+                'listen': no_value_msg,
+                'pg_hba': ['host all all all md5',
+                           f'host replication {os.environ["PATRONI_REPLICATION_USERNAME"]} all md5'],
+                'parameters': None,
+                'authentication': {'superuser': {'username': os.environ['PATRONI_SUPERUSER_USERNAME'],
+                                                 'password': os.environ['PATRONI_SUPERUSER_PASSWORD']},
+                                   'replication': {'username': os.environ["PATRONI_REPLICATION_USERNAME"],
+                                                   'password': os.environ['PATRONI_REPLICATION_PASSWORD']}},
+                'bin_dir': os.environ['PATRONI_POSTGRESQL_BIN_DIR']
+            }
+        }
+
+        # 2. Sample config
+
+        # Version-specific params
+        # 2.1 pg_version < 13, with dir creation
+        with patch('sys.argv', ['patroni.py', '--generate-sample-config', '/foo/bar.yml']),\
+             patch('subprocess.check_output', Mock(return_value=b"postgres (PostgreSQL) 9.4.3")),\
+             self.assertRaises(SystemExit) as e:
+            patroni_main()
+        self.assertEqual(e.exception.code, 0)
+        self.assertEqual(config, mock_config_dump.call_args[0][0])
+        mock_makedir.assert_called_once()
+
+        mock_makedir.reset_mock()
+        mock_config_dump.reset_mock()
+
+        # 2.2 pg_version >= 13
+        del config['bootstrap']['dcs']['postgresql']['parameters']['wal_keep_segments']
+        config['bootstrap']['dcs']['postgresql']['parameters']['wal_keep_size'] = '128MB'
+        config['bootstrap']['dcs']['postgresql']['use_pg_rewind'] = True
+        config['postgresql']['authentication']['rewind'] = {'username': os.environ['PATRONI_REWIND_USERNAME'],
+                                                            'password': no_value_msg}
+        config['postgresql']['pg_hba'] = ['host all all all scram-sha-256',
+                                          f'host replication {os.environ["PATRONI_REPLICATION_USERNAME"]} all scram-sha-256']
+        with patch('sys.argv', ['patroni.py', '--generate-sample-config', '/foo/bar.yml']),\
+             patch('subprocess.check_output', Mock(return_value=b"postgres (PostgreSQL) 15.2")),\
+             self.assertRaises(SystemExit) as e:
+            patroni_main()
+        self.assertEqual(e.exception.code, 0)
+        self.assertEqual(config, mock_config_dump.call_args[0][0])
+
+        mock_config_dump.reset_mock()
+
+        # 3. Config for a running PG13+ instance
+        del config['bootstrap']['dcs']['postgresql']['parameters']['wal_keep_size']
+
+        # values are taken from tests/__init__.py
+        config['scope'] = 'my_cluster'  # cluster_name
+        config['postgresql']['connect_address'] = 'foo:bar'
+        config['postgresql']['listen'] = '6.6.6.6:1984'
+        config['postgresql']['parameters'] = {'archive_command': 'my archive command'}
+        config['postgresql']['parameters']['hba_file'] = '/hba/file/path'
+        config['postgresql']['parameters']['ident_file'] = '/ident/file/path'
+        config['postgresql']['pg_ident'] = ['foo bar foobar', 'foobar bar foo']
+        config['bootstrap']['dcs']['postgresql']['parameters']['max_connections'] = 42
+        config['bootstrap']['dcs']['postgresql']['parameters']['max_locks_per_transaction'] = 73
+        config['bootstrap']['dcs']['postgresql']['parameters']['max_replication_slots'] = 21
+        config['bootstrap']['dcs']['postgresql']['parameters']['max_wal_senders'] = 37
+        config['bootstrap']['dcs']['postgresql']['parameters']['wal_level'] = 'replica'
+        config['postgresql']['data_dir'] = '/foo/bar/data'
+        config['postgresql']['bin_dir'] = '/bin/dir/from/running'
+        config['postgresql']['authentication']['superuser'] = {
+            'username': 'foobar',
+            'password': 'qwerty',
+            'channel_binding': 'prefer',
+            'gssencmode': 'prefer',
+            'sslmode': 'prefer'
+        }
+        config['postgresql']['authentication']['replication'] = {'username': no_value_msg, 'password': no_value_msg}
+
+        # Rewind is not configured for a running instance config
+        del config['bootstrap']['dcs']['postgresql']['use_pg_rewind']
+        del config['postgresql']['authentication']['rewind']
+
+        hba_content = '\n'.join(config['postgresql']['pg_hba'] + ['#host all all all md5'])
+        ident_content = '\n'.join(config['postgresql']['pg_ident'] + ['# something very interesting', '  '])
+        open_res = []
+        for _ in range(5):
+            open_res.extend([
+                mock_open(read_data=hba_content)(),
+                mock_open(read_data=ident_content)(),
+                mock_open(read_data='1984')(),
+                mock_open()()
+            ])
+
+        with patch('psutil.Process.__init__', Mock(return_value=None)),\
+             patch('psutil.Process.exe', Mock(return_value='/bin/dir/from/running/postgres')),\
+             patch('subprocess.check_output', Mock(return_value=b"postgres (PostgreSQL) 15.2")),\
+             patch('builtins.open', Mock(side_effect=open_res)):
+
+            # 3.1.1 pg_version >= 13
+            with patch('sys.argv', ['patroni.py', '--generate-config',
+                                    '--dsn', 'host=foo port=bar user=foobar password=qwerty']),\
+                 self.assertRaises(SystemExit) as e:
+                patroni_main()
+            self.assertEqual(e.exception.code, 0)
+            self.assertEqual(config, mock_config_dump.call_args[0][0])
+
+            mock_config_dump.reset_mock()
+
+            # 3.1.2 su auth params and connect host from env
+            os.environ['PGCHANNELBINDING'] =\
+                config['postgresql']['authentication']['superuser']['channel_binding'] = 'disable'
+            del config['postgresql']['authentication']['superuser']['gssencmode']
+            del config['postgresql']['authentication']['superuser']['sslmode']
+            config['postgresql']['authentication']['superuser']['username'] = os.environ['PGUSER']
+            config['postgresql']['authentication']['superuser']['password'] = os.environ['PGPASSWORD']
+            config['postgresql']['connect_address'] = f"{os.environ['PGHOST']}:1984"
+
+            with patch('sys.argv', ['patroni.py', '--generate-config']):
+
+                with self.assertRaises(SystemExit) as e:
+                    patroni_main()
+
+                self.assertEqual(config, mock_config_dump.call_args[0][0])
+
+                mock_config_dump.reset_mock()
+
+                # 3.1.5 Error while calling postgres --version
+                with patch('subprocess.check_output', Mock(side_effect=OSError)),\
+                     self.assertRaises(SystemExit) as e:
+                    patroni_main()
+                self.assertEqual(e.exception.code, 1)
+                self.assertIn('Failed to get postgres version:', mock_stderr.getvalue())
+
+                with patch('builtins.open', Mock(side_effect=[mock_open(read_data=hba_content)(),
+                                                              mock_open(read_data=ident_content)(),
+                                                              mock_open(read_data='')(),  # 3.2 empty postmaster.pid
+                                                              mock_open(read_data=hba_content)(),
+                                                              mock_open(read_data=ident_content)(),
+                                                              OSError,  # 3.3 Failed to open postmaster.pid
+                                                              mock_open(read_data=hba_content)(),
+                                                              mock_open(read_data=ident_content)(),
+                                                              mock_open(read_data='1984')()  # 3.4 Invalid postmaster pid
+                                                              ])):
+                    with self.assertRaises(SystemExit) as e:
+                        patroni_main()
+                    self.assertEqual(e.exception.code, 1)
+
+                    with self.assertRaises(SystemExit) as e:
+                        patroni_main()
+                    self.assertEqual(e.exception.code, 1)
+
+                    with patch('psutil.Process.__init__', Mock(return_value=None)),\
+                         patch('psutil.Process.exe', Mock(side_effect=psutil.NoSuchProcess(1984))),\
+                         self.assertRaises(SystemExit) as e:
+                        patroni_main()
+                    self.assertEqual(e.exception.code, 1)
+
+                # 3.5 Failed to open pg_hba
+                with patch('builtins.open', Mock(side_effect=OSError)),\
+                     self.assertRaises(SystemExit) as e:
+                    patroni_main()
+                self.assertEqual(e.exception.code, 1)
+
+                # 3.6 Failed to open pg_ident
+                with patch('builtins.open', Mock(side_effect=[mock_open(read_data=hba_content)(), OSError])),\
+                     self.assertRaises(SystemExit) as e:
+                    patroni_main()
+                self.assertEqual(e.exception.code, 1)
+
+                # 3.7 Failed PG connecttion
+                from . import psycopg
+                with patch('patroni.psycopg.connect', side_effect=psycopg.Error),\
+                     self.assertRaises(SystemExit) as e:
+                    patroni_main()
+                self.assertEqual(e.exception.code, 1)
+
+                # 3.8 Failed to get local IP
+                with patch('socket.socket.connect', Mock(side_effect=OSError)),\
+                     self.assertRaises(SystemExit) as e:
+                    patroni_main()
+                self.assertEqual(e.exception.code, 1)
 
     @patch('pkgutil.iter_importers', Mock(return_value=[MockFrozenImporter()]))
     @patch('sys.frozen', Mock(return_value=True), create=True)
